♻️ Refactor the schedule into separate component
