+++ src/components/station-detail/station-detail.js
@@ -1,19 +1,15 @@
-import React, { Component } from 'react';
-import { Fixed, Heading, Modal, Image, Flex, Box, Panel, Link, Text, Close, ButtonCircle, Hide } from 'rebass';
-import { Document, Page } from 'react-pdf/dist/entry.webpack';
+import React from 'react';
+import { Fixed, Modal, Flex, Box, Panel, Link, Text, ButtonCircle } from 'rebass';
 
 import ComingSoon from '../coming-soon/coming-soon';
+import StationSchedule from '../station-schedule/station-schedule';
 
 import { stationsByName } from '../../config/stations';
 import lines from '../../config/lines';
-import stationConfig from '../map/stationConfig';
 
 // This page does not work, currently.
 const getDeparturesUrl = stationCode =>
   `https://www.gotransit.com/en/trip-planning/scheduled-departures/station?stationCode=${stationCode}`;
-
-const pdfFilepath = station => 
-  `../../assets/pdfs/${stationsByName[station].line}/${stationsByName[station].miniScheduleFilename}`;
 
 const StationDetail = ({ station, onModalClick }) => (
   <div>
@@ -41,7 +37,7 @@ const StationDetail = ({ station, onModalClick }) => (
           bg='black'>
           <Flex alignItems='center'>
             <Box flex={1} py={1}>
-               <Text textAlign='center'>{stationConfig[station].name}</Text>
+               <Text textAlign='center'>{stationsByName[station].name}</Text>
             </Box>
             <Box>
               <ButtonCircle
@@ -61,41 +57,11 @@ const StationDetail = ({ station, onModalClick }) => (
           justifyContent="space-around"
         >
           <Box p={3}>
-            <Hide xs sm md lg>
-              <Document
-                file={pdfFilepath(station)}
-                >
-                <Page pageNumber={2} scale={1.9}/>
-              </Document>
-            </Hide>
-            <Hide xs sm md xl>
-              <Document
-                file={pdfFilepath(station)}
-                >
-                <Page pageNumber={2} scale={1.65}/>
-              </Document>
-            </Hide>
-            <Hide xs sm lg xl>
-              <Document
-                file={pdfFilepath(station)}
-                >
-                <Page pageNumber={2} scale={1.45}/>
-              </Document>
-            </Hide>
-            <Hide xs md lg xl>
-              <Document
-                file={pdfFilepath(station)}
-                >
-                <Page pageNumber={2} scale={1.15}/>
-              </Document>
-            </Hide>
-            <Hide sm md lg xl>
-              <Document
-                file={pdfFilepath(station)}
-                >
-                <Page pageNumber={2} scale={0.80}/>
-              </Document>
-            </Hide>
+            {
+              stationsByName[station].miniScheduleFilename
+              ? <StationSchedule station={station} />
+              : <ComingSoon />
+            }
           </Box>
         </Flex>
         <Panel.Footer
@@ -105,7 +71,7 @@ const StationDetail = ({ station, onModalClick }) => (
           <Flex>
             <Box flex={1}>
               <Link
-                href={stationConfig[station].fullScheduleUrl}
+                href={lines[stationsByName[station].line].fullScheduleUrl}
                 target="_blank"
               >
                 <Text textAlign='center'>Full Schedule</Text>
